==> notebooks/exploration/10_Feature_Generation.ipynb <==
import pandas as pd 
import seaborn as sns
import matplotlib.pyplot as plt
import numpy as np

data = pd.read_csv("../../data/processed/data.csv",
                   dtype = {
                            "expensa_id": "object",
                            "unidad_id": "object",
                            "propietario_id": "object",
                            "inquilino_id": "object",
                            "consorcio_id": "object",
                            "expensa_padre_id": "object",
                            "unidad_prop": "object",
                            "consorcio_id": "object",
                            "usuario_creador_id": "object",
                            "concepto_id": "object",
                            "mes": "object",
                            "anio": "object",
                            "unidad_denominacion":"object",
                            "consorcio_nombre": "object",
                            "version":"object",
                            "concepto_id": "object",
                            "unidad_tipo": "object",
                            "modo_pago": "object",
                            "forma_prorrateo": "object",
                            "tipo_prorrateo": "object",
                            "metodo_pago": "object",                       
                    },
                    parse_dates = ["fecha_vencimiento_1", "fecha_vencimiento_2"])
data.fecha_pago = pd.to_datetime(data.fecha_pago, errors="coerce")
data["fecha_expensa"] = data.anio + '-' + data.mes + '-01'
data.fecha_expensa = pd.to_datetime(data.fecha_expensa)
data["fecha_expensa_anterior"] = data.fecha_expensa - pd.DateOffset(months=1)
data["impago"] = data.fecha_pago.isna()
data["pago_tardio_1"] = (data.fecha_pago > data.fecha_vencimiento_1) & (data.fecha_pago <= data.fecha_vencimiento_2)
data["pago_tardio_2"] = (data.fecha_pago > data.fecha_vencimiento_2)
data["pago_tardio"] = data.pago_tardio_1 | data.pago_tardio_2
data.loc[data.metodo_pago.isna(),"metodo_pago"] = "Impago"

data.shape

## Agregado de Tamaño de Consorcio

consorcios_raw = pd.read_csv("../../data/raw/expensas_consorcios.csv",
                         sep = ";",
                         na_values = "NULL",
                         dtype = {
                            "id": "object",
                        })

unidades_raw = pd.read_csv("../../data/raw/expensas_unidades.csv",
                         sep = ";",
                         na_values = "NULL",
                         dtype = {
                            "id": "object",
                            "consorcio_id":"object"
                        })

consorcios = consorcios_raw.loc[consorcios_raw.deleted == 0, ["id"]]
unidades = unidades_raw.loc[unidades_raw.consorcio_id.isin(consorcios.id),:]
unidades = unidades.loc[unidades.deleted == 0,:]
unidades_consorcio = pd.merge(consorcios,
                              unidades,
                              left_on="id",
                              right_on="consorcio_id",
                              suffixes = ("","_unidad"))
unidad_por_consorcio = unidades_consorcio.groupby("consorcio_id").count()["id_unidad"].reset_index()

unidad_por_consorcio["log_tamanio_consorcio"] = np.round(np.log(unidad_por_consorcio.id_unidad))
unidad_por_consorcio.loc[unidad_por_consorcio.log_tamanio_consorcio <=1, "log_tamanio_consorcio"] = 1
unidad_por_consorcio.loc[unidad_por_consorcio.log_tamanio_consorcio >=5, "log_tamanio_consorcio"] = 5
unidad_por_consorcio["tamanio_consorcio"] = unidad_por_consorcio.log_tamanio_consorcio.map(str)
tamanios_consorcios = unidad_por_consorcio[["consorcio_id", "tamanio_consorcio"]]

data = pd.merge(data,
                tamanios_consorcios,
                left_on="consorcio_id",
                right_on="consorcio_id",
                how="left")

data.shape

data.columns

data = data[['expensa_id', 'unidad_id', 'fecha_vencimiento_1', 'fecha_vencimiento_2', 'monto', 'monto_total', 'int_dia', 'int_mes', 'mes', 'unidad_denominacion', 'unidad_prop', 'consorcio_id',
       'consorcio_nombre', 'version', 'fecha_pago', 'metodo_pago', 'fecha_expensa', 'fecha_expensa_anterior',
       'impago', 'pago_tardio_1', 'pago_tardio_2', 'pago_tardio', 'tamanio_consorcio']]

data.isna().sum()

## Cantidad de Impagos hasta la fecha

def get_cum_by_unidad(unidad_id):
    data_unidad = data.loc[data.unidad_id == unidad_id, ["unidad_id","fecha_expensa", "impago", "pago_tardio", "pago_tardio_1", "pago_tardio_2"]].sort_values("fecha_expensa")
    data_unidad["cum_impago"] = data_unidad.impago.cumsum()
    data_unidad["cum_pago_tardio"] = data_unidad.pago_tardio.cumsum()
    data_unidad["cum_pago_tardio_1"] = data_unidad.pago_tardio_1.cumsum()
    data_unidad["cum_pago_tardio_2"] = data_unidad.pago_tardio_2.cumsum()
    return data_unidad

dfs = list()
for unidad in data.unidad_id.unique():
    data_unidad = get_cum_by_unidad(unidad)
    dfs.append(data_unidad)

final = pd.concat(dfs)

final.shape

data = pd.merge(data,
                final[["unidad_id","fecha_expensa", "cum_impago", "cum_pago_tardio", "cum_pago_tardio_1", "cum_pago_tardio_2"]],
                left_on=["fecha_expensa_anterior","unidad_id"],
                right_on=["fecha_expensa", "unidad_id"],
                how = "left")

data["venc_1_dayofweek"] = data.fecha_vencimiento_1.dt.dayofweek.map(str)
data["venc_1_day"] = data.fecha_vencimiento_1.dt.day
data["venc_2_dayofweek"] = data.fecha_vencimiento_2.dt.dayofweek.map(str)
data["venc_2_day"] = data.fecha_vencimiento_2.dt.day
data["mes_vencimiento"] = data.fecha_vencimiento_1.dt.year.map(str) + '-' + data.fecha_vencimiento_1.dt.month.map(str) + '-01'
data.mes_vencimiento = pd.to_datetime(data.mes_vencimiento)
data["dia_inicio_mes_vencimiento"] = data.mes_vencimiento.dt.dayofweek

data_mes_anterior = data[["fecha_expensa_x", "unidad_id", "venc_1_dayofweek", "venc_1_day", "venc_2_dayofweek", "venc_2_day", "metodo_pago", "impago", "pago_tardio", "pago_tardio_1", "pago_tardio_2", "cum_impago", "cum_pago_tardio", "cum_pago_tardio_1", "cum_pago_tardio_2"]]

data = pd.merge(
    data,
    data_mes_anterior,
    left_on=["fecha_expensa_anterior","unidad_id"],
    right_on=["fecha_expensa_x", "unidad_id"],
    how = "left",
    suffixes = ("","_ant"))

data.to_csv("../../data/processed/data_2.csv", index=False)
