==> notebooks/exploration/12_Classification_6monthsago_Approach.ipynb <==
import pandas as pd 
import seaborn as sns
import matplotlib.pyplot as plt
import numpy as np

data = pd.read_csv("../../data/processed/data_2.csv",
                   dtype = {
                            "expensa_id": "object",
                            "unidad_id": "object",
                            "propietario_id": "object",
                            "inquilino_id": "object",
                            "consorcio_id": "object",
                            "expensa_padre_id": "object",
                            "unidad_prop": "object",
                            "consorcio_id": "object",
                            "usuario_creador_id": "object",
                            "concepto_id": "object",
                            "mes": "object",
                            "anio": "object",
                            "unidad_denominacion":"object",
                            "consorcio_nombre": "object",
                            "version":"object",
                            "concepto_id": "object",
                            "unidad_tipo": "object",
                            "modo_pago": "object",
                            "forma_prorrateo": "object",
                            "tipo_prorrateo": "object",
                            "metodo_pago": "object",    
                            "venc_1_dayofweek": "object",
                            "venc_2_dayofweek": "object",
                            "venc_1_dayofweek_ant": "object",
                            "venc_2_dayofweek_ant": "object",
                            "dia_inicio_mes_vencimiento": "object"
                    },
                    parse_dates = ["fecha_vencimiento_1", "fecha_vencimiento_2"])

data = data[['unidad_id', 'unidad_tipo', 'modo_pago', 'tamanio_consorcio',
             'fecha_expensa_x','mes_vencimiento','dia_inicio_mes_vencimiento',
             'fecha_vencimiento_1', 'venc_1_dayofweek', 'venc_1_day',
             'fecha_vencimiento_2', 'venc_2_dayofweek', 'venc_2_day',             
             'monto', 'monto_total', 'monto_parcial', 
             'int_dia', 'int_mes', 
             'cum_impago_ant', 'cum_pago_tardio_1_ant','cum_pago_tardio_2_ant',
             'mes',  
             'version',
             'metodo_pago',
             'impago', 'pago_tardio_1', 'pago_tardio_2',
             'fecha_expensa_anterior']]

data.columns = ['unidad_id', 'unidad_tipo', 'modo_pago', 'tamanio_consorcio',
                 'fecha_expensa','mes_vencimiento','dia_inicio_mes_vencimiento',
                 'fecha_vencimiento_1', 'fecha_vencimiento_1_dayofweek', 'fecha_vencimiento_1_daynumber',
                 'fecha_vencimiento_2', 'fecha_vencimiento_2_dayofweek', 'fecha_vencimiento_2_daynumber',             
                 'monto', 'monto_total', 'monto_parcial', 
                 'int_dia', 'int_mes', 
                 'cantidad_impagos_a_la_fecha', 'cantidad_pagos_tardios_1_a_la_fecha','cantidad_pagos_tardios_2_a_la_fecha',
                 'mes',  
                 'version',        
                 'metodo_pago',
                 'impago', 'pago_tardio_1', 'pago_tardio_2',
                 'fecha_expensa_anterior']

data["pago_a_tiempo"] = ~(data.impago | data.pago_tardio_1 | data.pago_tardio_2)

data = data.set_index(["unidad_id","fecha_expensa"])

data_mes_anterior = data[['modo_pago', 'dia_inicio_mes_vencimiento', 
                          'fecha_vencimiento_1_dayofweek', 'fecha_vencimiento_1_daynumber',
                          'fecha_vencimiento_2_dayofweek', 'fecha_vencimiento_2_daynumber', 
                          'metodo_pago',
                          'impago', 'pago_tardio_1', 'pago_tardio_2','pago_a_tiempo',
                          'fecha_expensa_anterior']]

data_1 = pd.merge(data,
         data_mes_anterior,
         left_on = ["unidad_id","fecha_expensa_anterior"],
         right_index = True,
         suffixes = ("", "_lag_1"))

data_1 = pd.merge(data_1,
         data_mes_anterior,
         left_on = ["unidad_id","fecha_expensa_anterior_lag_5"],
         right_index = True,
         suffixes = ("", "_lag_6"))

data_1.shape

data_1.columns

data_1 = data_1.reset_index()

X = data_1.drop(columns=['unidad_id', 'fecha_vencimiento_1', 'fecha_vencimiento_2','fecha_expensa','modo_pago','impago', 'pago_tardio_1', 'pago_tardio_2', 'pago_a_tiempo',
                       'fecha_expensa_anterior','fecha_expensa_anterior_lag_1', 'metodo_pago',
                      'fecha_expensa_anterior_lag_2', 'fecha_expensa_anterior_lag_3', 'fecha_expensa_anterior_lag_4',
                       'fecha_expensa_anterior_lag_5', 'fecha_expensa_anterior_lag_6'])

X = pd.get_dummies(X)

X.shape

X_train = X.loc[data_1.fecha_expensa < "2018-09-01", :]
X_test = X.loc[data_1.fecha_expensa >= "2018-09-01", :]
Y_train = ~target.loc[data_1.fecha_expensa < "2018-09-01", "pago_a_tiempo"]
Y_test = ~target.loc[data_1.fecha_expensa >= "2018-09-01", "pago_a_tiempo"]

from xgboost import XGBClassifier
from sklearn.model_selection import GridSearchCV

param_grid = {
    "max_depth": [3, 5, 7],
    "learning_rate": [0.1, 0.01],
    "n_estimators": [128,256,512],
}

xgb = GridSearchCV(
    estimator = XGBClassifier(verbosity=2, n_jobs=-1),
    param_grid = param_grid,   
    scoring = "f1",
    cv = 5,
    verbose = 2,
    return_train_score = True,
    n_jobs = -1
)

xgb.fit(X_train, Y_train)

xgb.score(X_train, Y_train)

xgb.score(X_test, Y_test)

importances_pago_tardio = pd.DataFrame({
    "variable": X.columns,
    "importance": xgb.best_estimator_.feature_importances_
})

importances_pago_tardio.sort_values("importance", ascending=False)

from sklearn.metrics import confusion_matrix, accuracy_score, precision_score, recall_score
confusion_matrix(Y_test, xgb.predict(X_test))

print("accuracy",accuracy_score(Y_train, xgb.predict(X_train)))
print("precision",precision_score(Y_train, xgb.predict(X_train)))
print("recall",recall_score(Y_train, xgb.predict(X_train)))

print("accuracy",accuracy_score(Y_test, xgb.predict(X_test)))
print("precision",precision_score(Y_test, xgb.predict(X_test)))
print("recall",recall_score(Y_test, xgb.predict(X_test)))

xgb.best_estimator_.save_model("../../models/xgb_paga_a_tiempo")

target = data_1.loc[:, ['impago', 'pago_tardio_1', 'pago_tardio_2','pago_a_tiempo']]

target["target"] = target["impago"].map(int).map(str) + target["pago_tardio_1"].map(int).map(str) +target["pago_tardio_2"].map(int).map(str)

target.loc[target.target == '000', "target"] = "A Tiempo"
target.loc[target.target == '100', "target"] = "Impago"
target.loc[target.target == '010', "target"] = "Tardio 1"
target.loc[target.target == '001', "target"] = "Tardio 2"
t

target
