==> notebooks/exploration/15_LSTM_Ordinal_Approach.ipynb <==
import pandas as pd 
import matplotlib.pyplot as plt
import numpy as np

data = pd.read_csv("../../data/processed/data_2.csv",
                   dtype = {
                            "expensa_id": "object",
                            "unidad_id": "object",
                            "propietario_id": "object",
                            "inquilino_id": "object",
                            "consorcio_id": "object",
                            "expensa_padre_id": "object",
                            "unidad_prop": "object",
                            "consorcio_id": "object",
                            "usuario_creador_id": "object",
                            "concepto_id": "object",
                            "mes": "object",
                            "anio": "object",
                            "unidad_denominacion":"object",
                            "consorcio_nombre": "object",
                            "version":"object",
                            "concepto_id": "object",
                            "unidad_tipo": "object",
                            "modo_pago": "object",
                            "forma_prorrateo": "object",
                            "tipo_prorrateo": "object",
                            "metodo_pago": "object",    
                            "venc_1_dayofweek": "object",
                            "venc_2_dayofweek": "object",
                            "venc_1_dayofweek_ant": "object",
                            "venc_2_dayofweek_ant": "object",
                            "dia_inicio_mes_vencimiento": "object"
                    },
                    parse_dates = ["fecha_vencimiento_1", "fecha_vencimiento_2"])

data = data[['unidad_id', 'unidad_tipo', 'modo_pago', 'tamanio_consorcio',
             'fecha_expensa_x','mes_vencimiento','dia_inicio_mes_vencimiento',
             'fecha_vencimiento_1', 'venc_1_dayofweek', 'venc_1_day',
             'fecha_vencimiento_2', 'venc_2_dayofweek', 'venc_2_day',             
             'monto', 'monto_total', 'monto_parcial', 
             'int_dia', 'int_mes', 
             'cum_impago_ant', 'cum_pago_tardio_1_ant','cum_pago_tardio_2_ant',
             'mes',  
             'version',
             'metodo_pago',
             'impago', 'pago_tardio_1', 'pago_tardio_2',
             'fecha_expensa_anterior']]

data.columns = ['unidad_id', 'unidad_tipo', 'modo_pago', 'tamanio_consorcio',
                 'fecha_expensa','mes_vencimiento','dia_inicio_mes_vencimiento',
                 'fecha_vencimiento_1', 'fecha_vencimiento_1_dayofweek', 'fecha_vencimiento_1_daynumber',
                 'fecha_vencimiento_2', 'fecha_vencimiento_2_dayofweek', 'fecha_vencimiento_2_daynumber',             
                 'monto', 'monto_total', 'monto_parcial', 
                 'int_dia', 'int_mes', 
                 'cantidad_impagos_a_la_fecha', 'cantidad_pagos_tardios_1_a_la_fecha','cantidad_pagos_tardios_2_a_la_fecha',
                 'mes',  
                 'version',        
                 'metodo_pago',
                 'impago', 'pago_tardio_1', 'pago_tardio_2',
                 'fecha_expensa_anterior']

data["target"] = data["impago"].map(int).map(str) + data["pago_tardio_1"].map(int).map(str) +data["pago_tardio_2"].map(int).map(str)

data.loc[data.target == '000', "target"] = "000"
data.loc[data.target == '100', "target"] = "100"
data.loc[data.target == '010', "target"] = "110"
data.loc[data.target == '001', "target"] = "111"
data = data.set_index(["unidad_id","fecha_expensa"])

data_mes_anterior = data[['target', 'fecha_expensa_anterior']]

data_lstm = data.reset_index()[["unidad_id","fecha_expensa","target", "fecha_expensa_anterior"]]

data_lstm = pd.merge(data_lstm,
         data_mes_anterior,
         left_on = ["unidad_id","fecha_expensa_anterior"],
         right_index = True,
         suffixes = ("", "_lag_1"))

for i in range(1,6):
    data_lstm = pd.merge(data_lstm,
                         data_mes_anterior,
                         left_on = ["unidad_id","fecha_expensa_anterior_lag_{}".format(i)],
                         right_index = True,
                         suffixes = ("", "_lag_{}".format(i+1)))

target_columns = ["fecha_expensa"] + [column for column in data_lstm.columns if column[0:6] == "target"]

data_lstm = data_lstm[target_columns]

data_lstm.shape

X = data_lstm.drop(columns="target")
Y = data_lstm[["fecha_expensa", "target"]]

X = X.loc[data_lstm.fecha_expensa < "2019-01-01",:]
Y = Y.loc[data_lstm.fecha_expensa < "2019-01-01",:]

print(X.shape)
print(Y.shape)

X.groupby([column for column in X.columns if column[0:6] == "target"]).count().reset_index().sort_values("fecha_expensa", ascending=False)

target_columns2 = [column for column in X.columns if column[0:6] == "target"]
X_train = X.loc[X.fecha_expensa < '2018-09-01', target_columns2]
X_test = X.loc[X.fecha_expensa >= '2018-09-01', target_columns2]
Y_train = Y.loc[Y.fecha_expensa < '2018-09-01', "target"]
Y_test =Y.loc[Y.fecha_expensa >= '2018-09-01', "target"]
Y_test_2 = Y.loc[Y.fecha_expensa >= '2018-09-01', "target"]
print(X_train.shape, X_test.shape)
print(Y_train.shape, Y_test.shape)

X_train = X_train.iloc[28:,:]
X_test = X_test.iloc[:-30,:]
Y_train = Y_train.iloc[28:]
Y_test = Y_test.iloc[:-30]
Y_test_2 = Y_test_2.iloc[:-30]

# /X_train = pd.get_dummies(X_train).values.reshape(X_train.shape[0],12,4)
# X_test = pd.get_dummies(X_test).values.reshape(X_test.shape[0],12,4)

def convert_to_ordinal_classes(df):
    for column in df.columns:
        df["N_" + column + "_PT1"] = df[column].str[0].map(int)
        df["N_" + column + "_PT2"] = df[column].str[1].map(int)
        df["N_" + column + "_NP"] = df[column].str[2].map(int)

    return df.iloc[:,6:].values.reshape(df.shape[0],6,3)
    

X_train = convert_to_ordinal_classes(X_train)
X_test = convert_to_ordinal_classes(X_test)

def convert_to_ordinal_classes_target(serie):
    df = pd.DataFrame({"target": serie})
    for column in df.columns:
        df["N_" + column + "_PT1"] = df[column].str[0].map(int)
        df["N_" + column + "_PT2"] = df[column].str[1].map(int)
        df["N_" + column + "_NP"] = df[column].str[2].map(int)

    return df.iloc[:,1:].values.reshape(df.shape[0],3)
Y_train = convert_to_ordinal_classes_target(Y_train)
Y_test = convert_to_ordinal_classes_target(Y_test)

# X_train = np.array(X_train.values).reshape(*X_train.shape)
# X_test = np.array(X_test.values).reshape(*X_test.shape)

# Y_train = np.array(pd.get_dummies(Y_train)[[0,1,2,3]].values).reshape(Y_train.shape[0],4)
# Y_test = np.array(pd.get_dummies(Y_test)[[0,1,2,3]].values).reshape(Y_test.shape[0],4)

print(X_train.shape, X_test.shape)
print(Y_train.shape, Y_test.shape)

from keras.layers.core import Dense, Dropout, Activation
from keras.models import Sequential
from keras.layers import LSTM, Embedding

lstm_model = Sequential()
lstm_model.add(LSTM(128, batch_input_shape=(100,6,3), dropout=0.0, recurrent_dropout=0.0, stateful=True, kernel_initializer='random_uniform'))
lstm_model.add(Dropout(0.3))
lstm_model.add(Dense(64,activation='relu'))
lstm_model.add(Dense(3,activation='sigmoid'))
lstm_model.compile(loss='categorical_crossentropy', optimizer="adam")

lstm_model.summary()

lstm_model.fit(X_train, Y_train, validation_data=(X_test, Y_test), epochs=40, verbose=2, batch_size=100,
                    shuffle=False)

from sklearn.metrics import confusion_matrix, accuracy_score, precision_score, recall_score
confusion_matrix(Y_test_2, lstm_model.predict_classes(X_test, batch_size=100))

pred = lstm_model.predict_classes(X_test, batch_size=100)

confusion_matrix(Y_test_2, pred)

def get_array_of_clases(array, threshold = 0.5):
    classes_array = np.ndarray((array.shape[0], 1))
    for i in range(0,array.shape[0]):
        for j in range(0,array.shape[1]):
            if array[i,j] < threshold:
                classes_array[i] = j
                break
            classes_array[i] = j + 1
    return classes_array

list(reversed(range(0, Y_test.shape[1])))

clases = get_array_of_clases(Y_test)

Y_test[678]

clases[678]

predictions = lstm_model.predict(X_test, batch_size=100)

get_array_of_clases(predictions)

Y_test[1:10]

from sklearn.metrics import confusion_matrix, accuracy_score, precision_score, recall_score
real = get_array_of_clases(Y_test)
predicted = get_array_of_clases(lstm_model.predict(X_test, batch_size=100), 0.90)
confusion_matrix(real, predicted)
