==> notebooks/exploration/18_Convolutional_Aproach.ipynb <==
import pandas as pd 
import matplotlib.pyplot as plt
import numpy as np

# Data Loading

data = pd.read_csv("../../data/processed/data.csv",
                   dtype = {
                            "expensa_id": "object",
                            "unidad_id": "object",
                            "propietario_id": "object",
                            "inquilino_id": "object",
                            "consorcio_id": "object",
                            "expensa_padre_id": "object",
                            "unidad_prop": "object",
                            "consorcio_id": "object",
                            "usuario_creador_id": "object",
                            "concepto_id": "object",
                            "mes": "object",
                            "anio": "object",
                            "unidad_denominacion":"object",
                            "consorcio_nombre": "object",
                            "version":"object",
                            "concepto_id": "object",
                            "unidad_tipo": "object",
                            "modo_pago": "object",
                            "forma_prorrateo": "object",
                            "tipo_prorrateo": "object",
                            "metodo_pago": "object",                       
                    },
                    parse_dates = ["fecha_vencimiento_1", "fecha_vencimiento_2"])
data.fecha_pago = pd.to_datetime(data.fecha_pago, errors="coerce")

data.shape

data.dtypes

# Generate Posible Target Variables

## Impago

data["target"] = ~((data.fecha_pago.dt.month == data.fecha_vencimiento_1.dt.month) & (data.fecha_pago.dt.year == data.fecha_vencimiento_1.dt.year)) 

data.target.sum()

data.target.sum()/data.target.count()

## Paga En el Mes?

data["fecha_expensa"] = data.anio + '-' + data.mes + '-01'
data.fecha_expensa = pd.to_datetime(data.fecha_expensa)
data["fecha_expensa_anterior"] = data.fecha_expensa - pd.DateOffset(months=1)

data_lstm = data[["unidad_id", "fecha_expensa", "target", "fecha_expensa_anterior"]]
data_mes_anterior = data.set_index(["unidad_id", "fecha_expensa"])

data_lstm = pd.merge(data_lstm,
         data_mes_anterior,
         left_on = ["unidad_id","fecha_expensa_anterior"],
         right_index = True,
         suffixes = ("", "_lag_1"))

for i in range(1,12):
    data_lstm = pd.merge(data_lstm,
                         data_mes_anterior,
                         left_on = ["unidad_id","fecha_expensa_anterior_lag_{}".format(i)],
                         right_index = True,
                         suffixes = ("", "_lag_{}".format(i+1)))

target_columns = ["fecha_expensa"] + [column for column in data_lstm.columns if column[0:6] == "target"]
data_lstm = data_lstm[target_columns]

X = data_lstm.drop(columns="target")
Y = data_lstm[["fecha_expensa", "target"]]

X = X.loc[data_lstm.fecha_expensa < "2019-01-01",:]
Y = Y.loc[data_lstm.fecha_expensa < "2019-01-01",:]

print(X.shape)
print(Y.shape)

target_columns2 = [column for column in X.columns if column[0:6] == "target"]
X_train = X.loc[X.fecha_expensa < '2018-09-01', target_columns2]
X_test = X.loc[X.fecha_expensa >= '2018-09-01', target_columns2]
Y_train = Y.loc[Y.fecha_expensa < '2018-09-01', "target"]
Y_test = Y.loc[Y.fecha_expensa >= '2018-09-01', "target"]
print(X_train.shape, X_test.shape)
print(Y_train.shape, Y_test.shape)

Y_train = Y_train.values.reshape(len(Y_train),1)

Y_test = Y_test.values.reshape(len(Y_test),1)

X_train_conv = X_train.values.reshape(*X_train.shape,1)
X_test_conv = X_test.values.reshape(*X_test.shape,1)

X_train.values.astype(float)

print(X_train_conv.shape, X_test_conv.shape)
print(Y_train.shape, Y_test.shape)

from keras.layers.core import Dense, Dropout, Activation, Flatten
from keras.models import Sequential
from keras.layers import Conv1D, Embedding, LSTM

conv_model = Sequential()
conv_model.add(Conv1D(filters=2, kernel_size = 7, input_shape=(12,1), padding= "same"))
conv_model.add(Flatten())
conv_model.add(Dense(12, activation="relu"))
conv_model.add(Dense(1, activation='sigmoid'))
conv_model.compile(loss='binary_crossentropy', optimizer="adam", metrics=['binary_accuracy'])

conv_model.summary()

conv_model.fit(X_train_conv, Y_train, epochs = 5, batch_size=128)

pred = conv_model.predict_classes(X_train_conv)

from sklearn.metrics import f1_score, precision_score, recall_score, accuracy_score
print("accuracy", accuracy_score(Y_train, pred))
print("f1", f1_score(Y_train, pred))
print("precision", precision_score(Y_train, pred))
print("recall", recall_score(Y_train, pred))

from sklearn.metrics import confusion_matrix

confusion_matrix(Y_train, pred)

pred_test = conv_model.predict_classes(X_test_conv)

print("accuracy", accuracy_score(Y_test, pred_test))
print("f1", f1_score(Y_test, pred_test))
print("precision", precision_score(Y_test, pred_test))
print("recall", recall_score(Y_test, pred_test))

## LSTM

X_train_lstm = X_train.values.reshape(*X_train.shape,1)
X_test_lstm = X_test.values.reshape(*X_test.shape,1)

X_train_lstm = X_train_lstm[9:,:,:]
X_test_lstm  = X_test_lstm[6:,:,:]
Y_train = Y_train[9:]
Y_test  = Y_test[6:]
print(X_train_lstm.shape, X_test_lstm.shape)
print(Y_train.shape, Y_test.shape)

from keras.layers.core import Dense, Dropout, Activation
from keras.models import Sequential
from keras.layers import LSTM, Embedding

lstm_model = Sequential()
lstm_model.add(LSTM(128, batch_input_shape=(10,6,1), dropout=0.0, recurrent_dropout=0.0, stateful=True, kernel_initializer='random_uniform'))
lstm_model.add(Dropout(0.3))
lstm_model.add(Dense(64,activation='relu'))
lstm_model.add(Dense(1,activation='softmax'))
lstm_model.compile(loss='binary_crossentropy', optimizer="adam")

lstm_model.summary()

lstm_model.fit(X_train_lstm, Y_train, epochs=2, verbose=3, batch_size=10, shuffle=True)

pred = lstm_model.predict_classes(X_train_lstm, batch_size=10)
print("accuracy", accuracy_score(Y_train, pred))
print("f1", f1_score(Y_train, pred))
print("precision", precision_score(Y_train, pred))
print("recall", recall_score(Y_train, pred))

pred = lstm_model.predict_classes(X_test_lstm, batch_size=10)
print("accuracy", accuracy_score(Y_test, pred))
print("f1", f1_score(Y_test, pred))
print("precision", precision_score(Y_test, pred))
print("recall", recall_score(Y_test, pred))
